feat(utils): created global utils module for both api and lit_processing with memoized and execute_once
