--- agr_literature_service/api/utils.py
@@ -1,7 +1,0 @@
-def execute_once(f):
-    def wrapper(*args, **kwargs):
-        if not wrapper.has_run:
-            wrapper.has_run = True
-            return f(*args, **kwargs)
-    wrapper.has_run = False
-    return wrapper
